Work done on API calls. Created container for Tiers.
